fix: reject --output without a value and reorganize tests
- --output の後に値がない場合を MissingOutputValue エラーにする
- FileTask のテストを config::tests から pipeline::tests に移動
- config_build_file_defers_name_resolution を
  config_build_single_file_without_output にリネーム
