+++ cli/cli/src/error.rs
@@ -15,6 +15,9 @@ pub enum CliError {
     #[error("Input path is neither a file nor directory: {0}")]
     InvalidInputPath(String),
 
+    #[error("--output requires a value")]
+    MissingOutputValue,
+
     #[error("--output cannot be used with directory input")]
     OutputWithDirectory,
